Return get response after API put request + fixed comment creation and display inconsistency
I changed the put request response to return a JsonResponse derived from the get_tutorials() function. This allows instant updates on the client side when making a promise, allowing data to be quickly recieved.

The purpose of this fix was to show the trash and reply buttons on the comments. This was difficult as the onclick function required data from the database, which couldn't be obtained through a function parameter.